pyproject.toml
[project]
name = "attestation"
version = "0.1.0"
description = ""
authors = [
    {name = "Rustam Gabdelislamov",email = "rustamgabdelislamov911@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "django-filter (>=25.2,<26.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-yasg (>=1.21.14,<2.0.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
